extracted loop closure code from relocalization code
